split the harness out
