+++ tests/common/harness.rs
@@ -1,0 +1,98 @@
+use crate::common::{mock_auth_service, mock_auth_service::AuthState};
+use clap::Parser;
+use kzg_ceremony_sequencer::{start_server, Options};
+use tempfile::{tempdir, TempDir};
+use tokio::sync::{broadcast, oneshot, Mutex, MutexGuard, OnceCell};
+
+fn test_options() -> Options {
+    let args: Vec<&str> = vec![
+        "kzg-ceremony-sequencer",
+        "--ceremony-sizes",
+        "4,3:8,3:16,3",
+        "--server",
+        "http://127.0.0.1:3000",
+        "--gh-token-url",
+        "http://127.0.0.1:3001/github/oauth/token",
+        "--gh-userinfo-url",
+        "http://127.0.0.1:3001/github/user",
+        "--gh-client-secret",
+        "INVALID",
+        "--gh-client-id",
+        "INVALID",
+        "--eth-rpc-url",
+        "INVALID",
+        "--eth-client-secret",
+        "INVALID",
+        "--eth-client-id",
+        "INVALID",
+        "--database-url",
+        "sqlite::memory:",
+    ];
+    Options::parse_from(args)
+}
+
+pub struct Harness {
+    pub options:     Options,
+    pub auth_state:  AuthState,
+    shutdown_sender: broadcast::Sender<()>,
+    /// Needed to keep the lock on the server port for the duration of a test.
+    #[allow(dead_code)]
+    lock:            MutexGuard<'static, ()>,
+    /// Needed to keep the temp directory alive throughout the test.
+    #[allow(dead_code)]
+    temp_dir:        TempDir,
+}
+
+impl Drop for Harness {
+    fn drop(&mut self) {
+        self.shutdown_sender.send(()).unwrap();
+    }
+}
+
+static SERVER_LOCK: OnceCell<Mutex<()>> = OnceCell::const_new();
+
+async fn server_lock() -> &'static Mutex<()> {
+    SERVER_LOCK.get_or_init(|| async { Mutex::new(()) }).await
+}
+
+pub async fn run_test_harness() -> Harness {
+    let lock = server_lock().await.lock().await;
+    let temp_dir = tempdir().unwrap();
+    let transcript = temp_dir.path().join("../../transcript.json");
+    let transcript_wip = temp_dir.path().join("transcript.json.next");
+    let mut options = test_options();
+    options.transcript_file = transcript;
+    options.transcript_in_progress_file = transcript_wip;
+    let server_options = options.clone();
+    let (shutdown_sender, mut app_shutdown_receiver) = broadcast::channel::<()>(1);
+    let mut auth_shutdown_receiver = shutdown_sender.subscribe();
+    let (app_start_sender, app_start_receiver) = oneshot::channel::<()>();
+    tokio::spawn(async move {
+        let server = start_server(server_options).await.unwrap();
+        app_start_sender.send(()).unwrap();
+        server
+            .with_graceful_shutdown(async move { app_shutdown_receiver.recv().await.unwrap() })
+            .await
+            .unwrap();
+    });
+    app_start_receiver.await.unwrap();
+    let (auth_start_sender, auth_start_receiver) = oneshot::channel::<()>();
+    let auth_state = AuthState::default();
+    let auth_state_for_server = auth_state.clone();
+    tokio::spawn(async move {
+        let server = mock_auth_service::start_server(auth_state_for_server);
+        auth_start_sender.send(()).unwrap();
+        server
+            .with_graceful_shutdown(async move { auth_shutdown_receiver.recv().await.unwrap() })
+            .await
+            .unwrap();
+    });
+    auth_start_receiver.await.unwrap();
+    Harness {
+        options: options.clone(),
+        auth_state,
+        shutdown_sender,
+        lock,
+        temp_dir,
+    }
+}
