+++ tests/common/mod.rs
@@ -1,0 +1,2 @@
+pub mod harness;
+pub mod mock_auth_service;
